Fix databases drop content when creating object
The content of the database is erased if it exists when a
KeyValueDatabase or a CollectionDatabase object is created.

The intent was to create an empty dictionary in the file when the
database did not exists, then the database class will be able to load
the file as a valid json object since loading an empty string as a JSON
object raises an error.

The problem was that forcing the empty dictionnary should be done only
if the file did not exists. Here it was done everytime the database
object was created.

The previous tests did not catch this because every tests uses the same
database object in a specific test and each test expect a clean
database.

This patch also adds a new tests to check this case.

+++ common_database.py
@@ -1,12 +1,14 @@
 
 import json
+import os
 
 class CommonDatabase():
 
     def __init__(self, db_path):
         self.db_path = db_path
-        with open(self.db_path, 'w') as file:
-            file.write("{}")
+        if not os.path.exists(self.db_path):
+            with open(self.db_path, 'w') as file:
+                file.write("{}")
 
     def save(self, content):
         with open(self.db_path, "w") as f:
